+++ CHANGELOG.md
@@ -12,7 +12,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `OCamlRuntime::releasing_runtime(&mut self, f: FnOnce() -> T)` releases the OCaml runtime, calls `f`, and then re-acquires the OCaml runtime. Maybe more complicated patterns should be supported, but for now I haven't given this much thought.
 - Support for unpacking OCaml polymorphic variants into Rust values.
 - Added `to_rust(cr: &OCamlRuntime)` method to `OCamlRooted<T>` values.
-- Added `keep_from_raw` method to root variables, an unsafe version of `keep_raw` that returns an `OCamlRooted<T>` instead of an `OCamlRawRooted`.
 
 ### Changed
 
@@ -24,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Rust functions that are exported to OCaml must now declare at least one argument.
 - Functions that are exported to OCaml now follow a caller-save convention. These functions now receive `OCamlRooted<T>`  parameters.
 - `to_rust()` method is now implemented directly into `OCaml<T>`, the `ToRust` trait is not required anymore.
+- `keep_raw()` method in root variables is now `unsafe` and returns an `OCamlRooted<T>`.
 
 ### Deprecated
 
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - Removed `ToRust` trait.
+- Removed `OCamlRawRooted` type.
 
 ### Fixed